Initial public commit
The first commit of this code was actually in a different repo back in
December 2019, but it was too tightly coupled with the temporary website
backend and database to be released publicly at the time. It has now
been significantly refactored into an independent package, but the core
logic hasn't really seen any changes since January.